Fix edge cases of patch event (time as int and provided not existing participants)

+++ backend/app/services/calendar_service.py
@@ -1,4 +1,5 @@
 from app.models.event import Event, EventParticipant
+from app.models.user import User
 from app.extensions import db
 from datetime import datetime
 
@@ -29,10 +30,8 @@ def create_event(data, owner_id):
 
     return new_event
 
-def patch_event(event, data):
-    allowed_fields = ['title', 'location', 'description', 'start_time', 'end_time']
-    updated_any = False
 
+def set_proposed_time(event, data):
     proposed_start = event.start_time
     proposed_end = event.end_time
 
@@ -42,9 +41,9 @@ def patch_event(event, data):
             try:
                 proposed_start = datetime.fromisoformat(value)
             except ValueError:
-                raise ValueError("Invalid format for start_time")
+                raise ValueError("INVALID_DATE: Invalid format for start_time")
         else:
-            proposed_start = value
+            raise TypeError(f"Invalid type for start_time: {type(value).__name__}")
 
     if 'end_time' in data:
         value = data['end_time']
@@ -52,9 +51,9 @@ def patch_event(event, data):
             try:
                 proposed_end = datetime.fromisoformat(value)
             except ValueError:
-                raise ValueError("Invalid format for end_time")
+                raise ValueError("INVALID_DATE: Invalid format for end_time")
         else:
-            proposed_end = value
+            raise TypeError(f"Invalid type for start_time: {type(value).__name__}")
 
     if proposed_start and proposed_end:
         check_start = proposed_start.replace(tzinfo=None) if proposed_start.tzinfo else proposed_start
@@ -62,6 +61,15 @@ def patch_event(event, data):
 
         if check_end <= check_start:
             raise ValueError("start_time must be before end_time.")
+
+    return proposed_start, proposed_end
+
+
+def patch_event(event, data):
+    allowed_fields = ['title', 'location', 'description', 'start_time', 'end_time']
+    updated_any = False
+
+    proposed_start, proposed_end = set_proposed_time(event, data)
 
     for field in allowed_fields:
         if field in data:
@@ -82,6 +90,13 @@ def patch_event(event, data):
         current_ids = {p.user_id for p in event.participant_links.all()}
 
         if new_ids != current_ids:
+            existing_users = db.session.query(User.id).filter(User.id.in_(new_ids)).all()
+            existing_user_ids = {user.id for user in existing_users}
+            invalid_ids = new_ids - existing_user_ids
+
+            if invalid_ids:
+                raise ValueError(f"NOT_FOUND Invalid user IDs: {', '.join(map(str, invalid_ids))}")
+
             for link in event.participant_links.all():
                 db.session.delete(link)
